fix: extract config to own file
